requirements.txt
Pillow==6.0.0
h5py==2.9.0
scipy==1.2.1
skimage==0.0
torch==1.0.1.post2
torchvision==0.2.2.post3
tqdm==4.31.1
